pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.15",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
    ]
